Built on Windows 10.
This is a Qt application which may require dlls in a Windows environment.  Installing Qt Creator and setting the path environmental variable to include (in my case) C:\Qt\5.15.2\mingw81_64\bin allowed this executable to run on my Windows 10 laptop as a stand alone program.  Building the program in Qt creator and running it through Creator does not require a change in the path environmental variable.